refactor(header): 목록 요소 컴포넌트 폴더, 파일명 변경 (#6)
추후 별도 route 폴더에서 재사용되는 컴포넌트로 components 폴더로 파일 위치 변경
컴포넌트명과 동일한 파일명으로 변경
